views questions

+++ question/views.py
@@ -1,3 +1,5 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib import messages
 from django.shortcuts import render, redirect, get_object_or_404, HttpResponse
 from django.views.generic import DetailView, CreateView
 from django.contrib.auth.decorators import login_required
@@ -5,14 +7,17 @@ from django.contrib.auth.decorators import login_required
 from .models import Question, Answer
 from .forms import QuestionForms, QuickQuestionForms, AnswerForms
 
+from user.models import Profile
+
 @login_required
-def create_question(request):
+def create_question(request,pk):
 
     if request.method == 'POST':
         form = QuestionForms(request.POST)
         if form.is_valid():
             object = form.save(commit=False)
             object.sender = request.user.profile
+            object.receiver = Profile.objects.get(id=pk)
             object.save()
             return redirect(object)
 
@@ -20,36 +25,44 @@ def create_question(request):
         form = QuestionForms()
 
     return render(request, "question/create.html", {"form": form})
-@login_required
-def del_question(request, pk):
-    # delete question only user who create it
-    objects = get_object_or_404(Question, pk=pk)
-    if objects.sender.pk == request.user.pk:
-        name_obj = objects.__str__()
-        objects.delete()
-        print(f'{name_obj}-- was deleted!')
-        #message
-        return redirect('home')
-    else:
-        print("You don't have permission")
-        return HttpResponse("<h1>You don't have permission</h1>")
 
-
-class DetailQuestionView(DetailView):
+class DetailQuestionView(LoginRequiredMixin, DetailView):
 
     model = Question
     template_name = 'question/detail.html'
 
-    # def get_context_data(self, **kwargs):
-    #     context = super().get_context_data(**kwargs)
-    #     context['form'] = QuickQuestionForms
-    #     return context
-    #
-    # def post(self, request, *args, **kwargs):
-    #     form = QuickQuestionForms(request.POST)
-    #     if form.is_valid():
-    #         print(self.pk_url_kwarg)
 
+@login_required
+def del_question(request, pk):
+    # delete question only user who create it
+    objects = get_object_or_404(Question, pk=pk)
+    if objects.sender.pk == request.user.pk and objects.can_edit():
+        name_obj = objects.__str__()
+        objects.delete()
+        messages.success(request, f'{name_obj}-- was deleted!')
+        return redirect('home')
+    else:
+        messages.error(request, "You don't have permission")
+        return redirect('home')
+
+@login_required
+def edit_question(request, pk):
+    obj = get_object_or_404(Question, pk=pk)
+    receiver = obj.receiver
+    if obj.can_edit() and request.user.pk == obj.sender.pk:
+        form = QuestionForms(instance=obj)
+        if request.method == 'POST':
+            form = QuestionForms(request.POST)
+            if form.is_valid():
+                object = form.save(commit=False)
+                object.sender = request.user.profile
+
+
+                object.receiver = receiver
+                object.save()
+                return redirect(object)
+        return render(request, "question/create.html", {"form": form})
+    return render(request, "question/info.html", {'info': 'Only the sender of the question can edit it.'})
 
 # ANSWER
 @login_required
